shipit_static_analysis: Refactor to support multiple reporting tools
