+++ src/shipit_static_analysis/shipit_static_analysis/clang/__init__.py
@@ -9,12 +9,17 @@ class ClangIssue(object):
     - path: Source file path relative to repository
     - line: Line where the issue begins
     - nb_lines: Number of lines affected by the issue
-    - mozreview_body: Build the text body published on MozReview
     '''
     def is_publishable(self):
         '''
         Is this issue publishable on Mozreview ?
         Should return a boolean
+        '''
+        raise NotImplementedError
+
+    def as_text(self):
+        '''
+        Build the text content for reporters
         '''
         raise NotImplementedError
