+++ src/shipit_static_analysis/shipit_static_analysis/config.py
@@ -18,6 +18,10 @@ logger = get_logger(__name__)
 class Settings(object):
     def __init__(self):
         self.config = None
+        self.app_channel = None
+
+    def setup(self, app_channel):
+        self.app_channel = app_channel
         self.download({
             'cpp_extensions': ['.cpp', '.c', '.h'],
         })
@@ -34,6 +38,7 @@ class Settings(object):
         Configuration is stored on mozilla central
         It has to be downloaded on each run
         '''
+        print('DOWNLOAD')
         assert isinstance(defaults, dict)
         assert self.config is None, \
             'Config already set.'
